设置打包字符编码

+++ common-ws-netty/src/main/java/com/awy/common/ws/netty/context/MessageManager.java
@@ -18,7 +18,7 @@ public class MessageManager {
         return Holder.manager;
     }
 
-    public Class<? extends Message> getMessage(int code){
+    public Class<? extends Message> getMessage(byte code){
         return msgRepository.get(code);
     }
